binary ninja: search for API using XDG desktop entry
ref #2376
