+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### New Features
 
 - add IDA v9.0 backend via idalib #2376 @williballenthin
+- locate Binary Ninja API using XDG Desktop Entries #2376 @williballenthin
 
 ### Breaking Changes
